# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users with their roles
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (name, email, status)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
  /security-policies/{id}:
    patch:
      operationId: updateSecurityPolicy
      tags: [security]
      summary: Update an app's security policy
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecurityPolicyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityPolicy"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
components:
  schemas:
...
    ErrorMessage:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    User:
      type: object
      required: [id, name, email, status, roles, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        status: { type: string, enum: [active, disabled] }
        entraObjectId: { type: string, nullable: true }
        roles:
          type: array
          items:
            type: object
            required: [assignmentId, roleId, roleName]
            properties:
              assignmentId: { type: integer }
              roleId: { type: integer }
              roleName: { type: string }
        createdAt: { type: string }
    UserInput:
      type: object
      required: [name, email]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, minLength: 3 }
        status: { type: string, enum: [active, disabled] }
        entraObjectId: { type: string }
        roleIds:
          type: array
          items: { type: integer }
    UserUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, minLength: 3 }
        status: { type: string, enum: [active, disabled] }
        entraObjectId: { type: string }
...
    SecurityPolicy:
      type: object
      required: [id, appId, appName, authMethod, mfaRequired, sessionTimeoutMinutes, recordLevelScope, fieldLevelRules, auditLogging, dataExportPolicy]
      properties:
        id: { type: integer }
        appId: { type: integer }
        appName: { type: string }
        authMethod: { type: string }
        mfaRequired: { type: string }
        sessionTimeoutMinutes: { type: integer }
        recordLevelScope: { type: string }
        fieldLevelRules: { type: string }
        auditLogging: { type: boolean }
        dataExportPolicy: { type: string }
    SecurityPolicyUpdate:
      type: object
      properties:
        authMethod: { type: string }
        mfaRequired: { type: string }
        sessionTimeoutMinutes: { type: integer }
        recordLevelScope: { type: string }
        fieldLevelRules: { type: string }
        auditLogging: { type: boolean }
        dataExportPolicy: { type: string }
...